Use business service in presentation layer
