+++ Presentation/pom.xml
@@ -12,10 +12,30 @@
 
     <artifactId>Presentation-Skeleton</artifactId>
 
+    <dependencyManagement>
+        <dependencies>
+            <dependency>
+                <groupId>hu.kalee</groupId>
+                <artifactId>Business-Skeleton</artifactId>
+                <version>${version}</version>
+            </dependency>
+        </dependencies>
+    </dependencyManagement>
+
     <dependencies>
+        <dependency>
+            <groupId>hu.kalee</groupId>
+            <artifactId>Business-Skeleton</artifactId>
+        </dependency>
+
         <dependency>
             <groupId>javax.servlet</groupId>
             <artifactId>jstl</artifactId>
+        </dependency>
+        <dependency>
+            <groupId>org.sonatype.sisu</groupId>
+            <artifactId>sisu-inject-bean</artifactId>
+            <version>1.4.2</version>
         </dependency>
 
         <!--<dependency>-->
