+++ Presentation/src/main/java/hu/kalee/skeleton/presentation/config/ServletInitializer.java
@@ -1,4 +1,4 @@
-package hu.kalee.skeleton.config;
+package hu.kalee.skeleton.presentation.config;
 
 import org.springframework.web.servlet.support.AbstractAnnotationConfigDispatcherServletInitializer;
