+++ Presentation/src/main/java/hu/kalee/skeleton/presentation/controller/HomeController.java
@@ -1,4 +1,4 @@
-package hu.kalee.skeleton.controller;
+package hu.kalee.skeleton.presentation.controller;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.RequestMapping;
