+++ pom.xml
@@ -10,6 +10,7 @@
     <version>0.1.0-SNAPSHOT</version>
     <modules>
         <module>Presentation</module>
+        <module>Business</module>
     </modules>
     <name>Spring MVC Skeleton</name>
